requirements.txt
beautifulsoup4==4.13.4
Flask==3.1.0
langchain_community==0.3.23
numpy==2.2.5
openai==1.78.0
pandas==2.2.3
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
sentence_transformers==4.1.0
torch==2.7.0
transformers==4.51.3
